added the word "becoming"
on line 214... before: Whether by accident, or on purpose, a smart contract can be created such that it runs forever when a node attempts to validate it, effectively a denial of service attack. 

after: Whether by accident, or on purpose, a smart contract can be created such that it runs forever when a node attempts to validate it, effectively becoming a denial of service attack.