Keys and Addresses - Typos and Omissions
The words "later" and "former" are used in the fourth paragraph of the section entitled "Introduction". Although later is synonymous to latter, latter is more often used in this circumstance. This is just a personal observation, I would not call this an "error".

The sentence "If the result within the valid range, we have a suitable private key" is missing a verb. It should read "If the result is within the valid range, we have a suitable private key" instead.

The sentence "(like the pseudo-random rand() function is most programming languages)" contains a typo. The word "is" does not work in this context and should be replaced with "in".