Mist was deprecated
I removed mist from the list of popular wallets.

I edited the mist section - added Deprecated to the title, changed present tense to past tense, and addeed a deprecated note and suggestion to use other wallets. 

I decided not to delete the mist section because it is important to know about historically.